discover bug in init of partition

+++ rgr/constants/types.pyx
@@ -1,7 +1,7 @@
 import numpy as np
 
 DTYPE_ADJ = np.int8
-DTYPE_STD = np.int16
+DTYPE_STD = np.int32
 DTYPE_IDX = np.uint32
 DTYPE_FLOAT = np.float32
